[fix] 색상 입히는 메소드 수정
- 원활한 로직을 위해 Color 타입 프로퍼티 Int -> Double로 수정

+++ DrawingApp/DrawingApp/Model/Properties/Color.swift
@@ -1,9 +1,9 @@
 struct Color {
-    var red:Int
-    var green:Int
-    var blue:Int
+    var red: Double
+    var green: Double
+    var blue: Double
     
-    init(red: Int, green: Int, blue: Int) {
+    init(red: Double, green: Double, blue: Double) {
         self.red = red
         self.green = green
         self.blue = blue
